Split PrimeTower class in (placeholder) subclasses
We now have specific classes for normal and interleaved prime towers, so
that it will become easier to fine-tune their behavior without breaking
each other.
