+++ src/FffPolygonGenerator.cpp
@@ -49,7 +49,7 @@
 #include "utils/ThreadPool.h"
 #include "utils/gettime.h"
 #include "utils/math.h"
-#include "PrimeTower.h"
+#include "PrimeTower/PrimeTower.h"
 #include "geometry/OpenPolyline.h"
 #include "utils/Simplify.h"
 // clang-format on
