+++ src/PrimeTower/PrimeTowerNormal.cpp
@@ -1,0 +1,14 @@
+// Copyright (c) 2024 UltiMaker
+// CuraEngine is released under the terms of the AGPLv3 or higher.
+
+#include "PrimeTower/PrimeTowerNormal.h"
+
+namespace cura
+{
+
+PrimeTowerNormal::PrimeTowerNormal(SliceDataStorage& storage, size_t extruder_count)
+    : PrimeTower(storage, extruder_count)
+{
+}
+
+} // namespace cura
